+++ src/dagster_ray/core/run_launcher.py
@@ -110,11 +110,11 @@ class RayRunLauncher(RunLauncher, ConfigurableClass):
 
     @functools.cached_property
     def client(self) -> JobSubmissionClient:
-        '''
+        """
         Cache the client to avoid creating multiple connections.
         When using ray:// addresses, JobSubmissionClient internally calls ray.init()
         which fails if already connected.
-        '''
+        """
         from ray.job_submission import JobSubmissionClient
 
         return JobSubmissionClient(self.address, metadata=self.metadata, headers=self.headers, cookies=self.cookies)